+++ CHANGELOG.md
@@ -8,9 +8,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Added support for reversed ranges (#834)
+- Added support for parsing formulas with all whitespace characters (controlled by config param `ignoreWhiteSpace`). (#898)
 
 ### Changed
-- Removed `gpu.js` dependency and its use. (#812)
+- **Breaking change**: Removed `gpu.js` dependency and its use. (#812)
+
+### Fixed
+- Fixed RATE function not converging for some inputs. (#905)
 
 ## [1.3.1] - 2022-01-11